+++ src/utils/index.ts
@@ -73,5 +73,6 @@ export * from './object-sanitize';
 export * from './permission-templates';
 export * from './smart-scope';
 export * from './service-act-reasons';
+export * from './same-as';
 export * from './activation-request';
 export * from './vp-token';